Add variable indices context to variable bridges
As the variable bridges substitute user (bridged) variables into
solver variables, the solver and the bridges that are created by this bridge are
using the solver variables while the user and other bridges are using
the user variables.

Before this commit, the solver variables where substituted for the user
variables even when the call was coming from a bridge that was created
(directly or indirectly) by the bridge that created these variables.
This causes issues if the call was for getting a function which should
then be converted into a VectorOfVariables as even if it is a
VectorOfVariables in terms of the solver variables, it may not be a
VectorOfVariables in terms of the user variables.