Use a Card struct when writing MPS files (#1173)
This fixes a bug, because stricter MPS readers like the one in Cbc
distinguish between the variable name "x" and " x" with a leading
space. This caused a segfault in a model sent to me by a user.

This also changes the logic, so that binary variables only get
associated bounds if the bounds are tigher than [0, 1], which Cbc
also complained about.