WIP: begin support for vector-valued objective functions
This is a re-hash of https://github.com/jump-dev/MathOptInterface.jl/pull/968

The last PR lost steam because of the discussion in https://github.com/jump-dev/JuMP.jl/issues/2099

My sense is that this is still the right API approach at the MOI-level, but
that it's the wrong approach at the JuMP level. I'm working on a PR to JuMP,
so let's hold off comments until things are working fully and we can discuss
the full API, rather than the pros and cons of treating multi-criteria as
vector optimization.

I have a proof-of-concept solver at https://github.com/odow/MOO.jl
and I'll also update Gurobi.jl to support the new API.