[Utilities] add with_cache_type to instantiate
This is useful when creating optimizers which do not support the
incremental interface, or which only partially implement it.

This change is motivated by some recent examples in which it was
necessary for the user to hard-code the caching logic to work-around
the lack of deletion in a solver, and the caching logic is confusing
to even advanced users. See:

 * https://discourse.julialang.org/t/jump-scip-error/93958
 * https://github.com/jump-dev/MultiObjectiveAlgorithms.jl/pull/46