Fix JET errors in bridge_optimizer.jl
This one deserved a separate commit because the changes are a little
subtle.

The various Bridge submodules define EmptyMap as a fallback for when
an AbstractBridgeOptimizer doesn't have an associated set of arcs
(like a Variable.SingleBridgeOptimizer). But this makes all calls
of .bridges(b) type unstable. And sometimes we really do want either,
but sometimes (like the ones annnotated here), we want only the explicit
Map.

For MOI v2, we could consider refactoring this code to remove EmptyMap
in favor of a more explicit fallback like nothing.